pyproject.toml
[project]
name = "seating"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "bcrypt>=4.0.0",
    "cryptography>=41.0.0",
    "email-validator>=2.0.0",
    "fastapi>=0.104.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.5.0",
    "pydantic[email]>=2.5.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.24.0",
]
